[Move] LCOrdeerPuzzleManager위치 변경
