+++ LastCanary/Content/_LastCanary/Dev/InGameTestLevel.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f5e67ef8d724a522cfcc6296de52a2b8515919af4c399e66ada1e8d8d515c0ea
-size 123032
+oid sha256:f4d5c10a5f14def16fcba10397d578cfdf6ff54a759a09729ef73bd1921a6c74
+size 123007
